code refactor and checking youtube video download speed

+++ app/src/main/java/com/example/ytsample/entities/ProgressState.kt
@@ -1,4 +1,4 @@
 package com.example.ytsample.entities
 
-class ProgressState(var percent:Int?,var length:Long?,var onProgress:Long?,var isFinished:Boolean?,var id:String) {
+class ProgressState(var percent:Int?,var length:Float?,var onProgress:Float?,var isFinished:Boolean?,var id:String) {
 }
